fixed rating count number of stars selected match rating count

+++ app/src/main/java/com/sable/businesslistingapi/RetrofitArrayApi.java
@@ -68,7 +68,7 @@ public interface RetrofitArrayApi {
             @Part("zip") String zip,
             @Part("latitude") Double latitude,
             @Part("longitude") Double longitude,
-            @Part("ratings") Integer ratings,
+            @Part("ratings") Float ratings,
             @Part("phone") String phone,
             @Part("email") String email,
             @Part("website") String website,
